+++ client/src/components/ServiceList.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import ServiceThumbnail from "./ServiceThumbnail.jsx";
+
+const ServiceList = (props) => (
+  <div>
+    {props.services.map(vet => <ServiceThumbnail service={service} key={service.id} />)}
+  </div>
+)
